Add files via upload
In home work, In code  I will change the value about sample rate to 8000, 16000, 22.050.

When changing the sample rates to [8000, 16000, 22050] Hz, we will notice a direct impact on the results provided by the audio classification system using YAMNet. This effect is related to how the model analyzes the data based on the sampling rate available to it.

sample rate 8000 Hz:
When using a sampling rate of 8000 Hz, the frequency range available to the YAMNet model will be reduced. This may result in the loss of some fine details in the audio and reduce classification accuracy.

 sample rate16000 Hz (default rate):
This is the rate at which YAMNet was trained, so the results here are expected to be the most accurate. This rate can provide the best balance between classification accuracy and resource consumption.

sample rate 22.050 Hz:
When using a 22050 Hz sampling rate, the frequency range available to the YAMNet model will be increased, which may increase the classification accuracy at high frequencies, but may also increase the load on the system.